requirements.txt
chromadb==0.4.22
langchain==0.1.6
langchain-openai==0.0.6
langchain-text-splitters==0.0.1
unstructured==0.12.4
python-docx==1.1.0
trafilatura==1.6.4
openai==1.12.0
python-dotenv==1.0.1
tiktoken==0.5.2
